:hammer: [Role&Permission] apply role and permissions on all actions

+++ app/Http/Controllers/CategoryController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Enums\PermissionEnum;
 use App\Models\Category;
 use App\Http\Requests\Category\StoreCategoryRequest;
 use App\Http\Requests\Category\UpdateCategoryRequest;
@@ -11,6 +12,7 @@ class CategoryController extends Controller
 {
     public function index()
     {
+        abort_unless(auth()->user()->can(PermissionEnum::READ_CATEGORIES), 403);
         $categories = Category::where("user_id", auth()->id())->count();
 
         return view('categories.index', [
@@ -20,11 +22,13 @@ class CategoryController extends Controller
 
     public function create()
     {
+        abort_unless(auth()->user()->can(PermissionEnum::CREATE_CATEGORIES), 403);
         return view('categories.create');
     }
 
     public function store(StoreCategoryRequest $request)
     {
+        abort_unless(auth()->user()->can(PermissionEnum::CREATE_CATEGORIES), 403);
         Category::create([
             "user_id"=>auth()->id(),
             "name" => $request->name,
@@ -38,6 +42,7 @@ class CategoryController extends Controller
 
     public function show(Category $category)
     {
+        abort_unless(auth()->user()->can(PermissionEnum::READ_CATEGORIES), 403);
         return view('categories.show', [
             'category' => $category
         ]);
@@ -45,6 +50,7 @@ class CategoryController extends Controller
 
     public function edit(Category $category)
     {
+        abort_unless(auth()->user()->can(PermissionEnum::UPDATE_CATEGORIES), 403);
         return view('categories.edit', [
             'category' => $category
         ]);
@@ -52,6 +58,7 @@ class CategoryController extends Controller
 
     public function update(UpdateCategoryRequest $request, Category $category)
     {
+        abort_unless(auth()->user()->can(PermissionEnum::UPDATE_CATEGORIES), 403);
         $category->update([
             "name" => $request->name,
             "slug" => Str::slug($request->name)
@@ -64,6 +71,7 @@ class CategoryController extends Controller
 
     public function destroy(Category $category)
     {
+        abort_unless(auth()->user()->can(PermissionEnum::DELETE_CATEGORIES), 403);
         $category->delete();
 
         return redirect()
